Store reverse path into waypoint
This makes the algorithm run a bit faster and makes the code simpler. Instead of relying only on the is_reversed field and checking that field every time we need to look up the endpoints of the element, we now simply swap the endpoints if the element needs to be reversed.

Contributes to issue CURA-329.